requirements.txt
django==5.0
djangorestframework==3.14.0
coverage==7.4.0
Pillow==10.1.0
requests==2.31.0
psycopg2==2.9.9